requirements.txt
flask==2.0.1
tensorflow==2.15.0
pillow==9.0.0
numpy==1.21.5
gunicorn==20.1.0 